+++ Source/Applications/Desktop/VodovozViewModels/Counterparties/ClientClassification/CounterpartyClassificationCalculationEmailSettingsViewModel.StartClassificationCalculationEventArgs.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Vodovoz.ViewModels.Counterparties.ClientClassification
+{
+	public partial class CounterpartyClassificationCalculationEmailSettingsViewModel
+	{
+		public class StartClassificationCalculationEventArgs : EventArgs
+		{
+			public string CurrentUserEmail { get; }
+			public string AdditionalEmail { get; }
+
+			public StartClassificationCalculationEventArgs(string currentUserEmail, string additionalEmail)
+			{
+				CurrentUserEmail = currentUserEmail;
+				AdditionalEmail = additionalEmail;
+			}
+		}
+	}
+}
